Revamped the ruby SDK structure and added accounts
