+++ ruby/lib/onvo/accounts.rb
@@ -1,12 +1,14 @@
 # frozen_string_literal: true
 
+require_relative '../resource_template'
+
 # Account endpoints
-module Accounts
-  def get_accounts
+class Accounts < ResourceTemplate
+  def list
     base_get('/accounts')
   end
 
-  def get_account_by_id(id)
+  def get(id)
     base_get("/accounts/#{id}")
   end
 end
