+++ ruby/lib/onvo/dashboard/questions.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'uri'
+
 # Dashboard Question endpoints
 module Questions
   def get_dashboard_questions_by_id(dashboard_id)
